requirements.txt
Flask>=3.0.0
pypdf>=3.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0

